refactor(export): split export_elections.py into an export/ package
Move the ~30 helper functions out of the 1,562-line script into focused
modules under scripts/export/: naming (slugs/names/value normalisation),
payload (pf-results-v4 result building), serialize (JSON writers), manifest
(map-modes.json building) and legacy (supplemental legacy elections).
export_elections.py keeps parse_args + main as the CLI orchestrator and
re-exports the moved names (tests import them from export_elections).

The OTHERS_PARTY_ID global moves into payload with a set_others_party_id()
setter that main calls at startup, so no module carries shared mutable state
implicitly.

Pure relocation: --current-simulation output and the full general-export tree
(16 files) are byte-identical to before; 246 tests pass.
